% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_restrictions.R
\name{remove_tag}
\alias{remove_tag}
\title{Remove cell position tags}
\usage{
remove_tag(dat, ...)
}
\arguments{
\item{dat}{\code{[data.frame]}}

\item{...}{Currently not used}
}
\value{
\code{dat} with tag removed
}
\description{
Remove cell position tags
}
\seealso{
Other tagging: \code{\link{tag_missing}},
  \code{\link{tagged_values}}
}
