% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_restrictions.R
\name{tag_missing}
\alias{tag_missing}
\title{Tag currently missing elements of a data.frame}
\usage{
tag_missing(dat, ...)
}
\arguments{
\item{dat}{\code{[data.frame]} to be tagged}

\item{...}{Currently not used.}
}
\value{
\code{dat}, tagged for missing values.
}
\description{
Attach an attribute that marks which cells are empty (NA).
}
\seealso{
Other tagging: \code{\link{remove_tag}},
  \code{\link{tagged_values}}
}
