% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_restrictions.R
\name{tagged_values}
\alias{tagged_values}
\title{Retrieve tagged cell positions}
\usage{
tagged_values(dat, ...)
}
\arguments{
\item{dat}{\code{[data.frame]}}

\item{...}{Currently not used}
}
\value{
A \code{logical} matrix, or \code{NULL}
}
\description{
Retrieve tagged cell positions
}
\seealso{
Other tagging: \code{\link{remove_tag}},
  \code{\link{tag_missing}}
}
